Add possibility to extend search sets by field types
